add copy to cliboard to it's own component
