+++ ui/components/UserGroupsTable.tsx
@@ -1,54 +1,14 @@
 import * as React from "react";
-import { useCallback, useState } from "react";
 import styled from "styled-components";
+import CopyToClipboard from "./CopyToCliboard";
 import DataTable, { Field } from "./DataTable";
 import Flex from "./Flex";
-import Icon, { IconType } from "./Icon";
 import MessageBox from "./MessageBox";
 import Text from "./Text";
 
 type Props = {
   className?: string;
   rows: string[];
-};
-
-const PointerIcon = styled(Icon)`
-  cursor: pointer;
-  svg {
-    fill: ${(props) => props.theme.colors.neutral30};
-  }
-  &.copied > svg {
-    fill: ${(props) => props.theme.colors.primary10};
-  }
-`;
-
-const CopyToClipboard = ({
-  value,
-  className,
-}: {
-  value: string;
-  className?: string;
-}) => {
-  const [copied, setCopied] = useState(false);
-  const handleCopy = useCallback(() => {
-    navigator.clipboard.writeText(value).then(() => {
-      setCopied(true);
-      setTimeout(() => {
-        setCopied(false);
-      }, 3000);
-    });
-  }, [value]);
-
-  return (
-    <Text onClick={handleCopy} data-testid="github-code" className={className}>
-      <span className="CopyText">{value}</span>
-      <PointerIcon
-        type={copied ? IconType.CheckMark : IconType.FileCopyIcon}
-        className={copied ? "copied" : ""}
-        size="medium"
-      />
-    </Text>
-  );
 };
 
 function UserGroupsTable({ className, rows }: Props) {
@@ -81,7 +41,7 @@ export default styled(UserGroupsTable).attrs({
   className: UserGroupsTable.name,
 })`
   .CopyToClipboard {
-    display: inline-flex;
+    display: flex;
     justify-content: center;
     align-items: center;
   }
